+++ packages/graphql-types/frontend/graphql.ts
@@ -278,11 +278,11 @@ export type DeploymentFrequencyMetric = {
   /** The columns for the chart */
   columns: Array<Scalars['DateTime']['output']>;
   /** The amount of deployments for the current period */
-  currentAmount: Scalars['Int']['output'];
+  currentAmount: Scalars['BigInt']['output'];
   /** The amounts over time for the chart */
   data: Array<Scalars['BigInt']['output']>;
   /** The number of deployments before the current period */
-  previousAmount: Scalars['Int']['output'];
+  previousAmount: Scalars['BigInt']['output'];
 };
 
 export type DeploymentSettings = {
@@ -469,12 +469,12 @@ export type LeadTimeMetric = {
   change: Scalars['Float']['output'];
   /** The columns for the chart */
   columns: Array<Scalars['DateTime']['output']>;
-  /** The lead time for the current period */
-  currentAmount: Scalars['Int']['output'];
+  /** The lead time in milliseconds for the current period */
+  currentAmount: Scalars['BigInt']['output'];
   /** The amounts over time for the chart */
   data: Array<Scalars['BigInt']['output']>;
-  /** The lead time before the current period */
-  previousAmount: Scalars['Int']['output'];
+  /** The lead time in milliseconds before the current period */
+  previousAmount: Scalars['BigInt']['output'];
 };
 
 export type LoginToStripeInput = {
@@ -499,11 +499,11 @@ export type MeanTimeToRecoverMetric = {
   /** The columns for the chart */
   columns: Array<Scalars['DateTime']['output']>;
   /** The mean time to recover in milliseconds for the current period */
-  currentAmount: Scalars['Int']['output'];
+  currentAmount: Scalars['BigInt']['output'];
   /** The amounts over time for the chart */
   data: Array<Scalars['BigInt']['output']>;
   /** The mean time to recover in milliseconds before the current period */
-  previousAmount: Scalars['Int']['output'];
+  previousAmount: Scalars['BigInt']['output'];
 };
 
 export type Metrics = {
